解决 erupt-cloud-server不兼容高版本 jdk 的问题
