+++ erupt-cloud/erupt-cloud-server/src/main/java/xyz/erupt/cloud/server/model/CloudNode.java
@@ -10,8 +10,6 @@ import xyz.erupt.annotation.EruptField;
 import xyz.erupt.annotation.EruptI18n;
 import xyz.erupt.annotation.constant.AnnotationConst;
 import xyz.erupt.annotation.expr.ExprBool;
-import xyz.erupt.annotation.fun.DataProxy;
-import xyz.erupt.annotation.fun.TagsFetchHandler;
 import xyz.erupt.annotation.sub_erupt.Layout;
 import xyz.erupt.annotation.sub_erupt.RowOperation;
 import xyz.erupt.annotation.sub_erupt.Tpl;
@@ -24,23 +22,10 @@ import xyz.erupt.annotation.sub_field.sub_edit.CodeEditorType;
 import xyz.erupt.annotation.sub_field.sub_edit.Search;
 import xyz.erupt.annotation.sub_field.sub_edit.TagsType;
 import xyz.erupt.cloud.server.base.CloudServerConst;
-import xyz.erupt.cloud.server.node.MetaNode;
-import xyz.erupt.cloud.server.node.NodeManager;
-import xyz.erupt.core.config.GsonFactory;
-import xyz.erupt.core.util.Erupts;
-import xyz.erupt.jpa.dao.EruptDao;
 import xyz.erupt.jpa.model.MetaModelUpdateVo;
-import xyz.erupt.linq.lambda.LambdaSee;
-import xyz.erupt.tpl.engine.EngineConst;
 import xyz.erupt.upms.handler.ViaMenuValueCtrl;
 
-import javax.annotation.Resource;
 import javax.persistence.*;
-import java.util.Collection;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.function.Function;
 
 /**
  * @author YuePeng
@@ -52,7 +37,7 @@ import java.util.function.Function;
 @Entity
 @Table(name = "e_cloud_node")
 @Erupt(
-        name = "节点配置", dataProxy = CloudNode.class,
+        name = "节点配置", dataProxy = CloudNodeProcess.class,
         rowOperation = @RowOperation(
                 title = "查看令牌", icon = "fa fa-shield", mode = RowOperation.Mode.SINGLE,
                 show = @ExprBool(exprHandler = ViaMenuValueCtrl.class, params = CloudServerConst.CLOUD_ACCESS_TOKEN_PERMISSION),
@@ -61,11 +46,7 @@ import java.util.function.Function;
 )
 @Component
 @EruptI18n
-public class CloudNode extends MetaModelUpdateVo implements DataProxy<CloudNode>, TagsFetchHandler, Tpl.TplHandler {
-
-    public static final String NODE_NAME = LambdaSee.field(CloudNode::getNodeName);
-
-    public static final String ACCESS_TOKEN = LambdaSee.field(CloudNode::getAccessToken);
+public class CloudNode extends MetaModelUpdateVo {
 
     @Column(unique = true)
     @EruptField(
@@ -110,7 +91,7 @@ public class CloudNode extends MetaModelUpdateVo implements DataProxy<CloudNode>
     @Transient
     @EruptField(
             views = @View(title = "实例数", className = "text-center", width = "70px"
-                    , tpl = @Tpl(path = "/tpl/node-instance.ftl", width = "400px", tplHandler = CloudNode.class)
+                    , tpl = @Tpl(path = "/tpl/node-instance.ftl", width = "400px", tplHandler = CloudNodeProcess.class)
             )
     )
     private Integer instanceNum;
@@ -124,7 +105,7 @@ public class CloudNode extends MetaModelUpdateVo implements DataProxy<CloudNode>
     @EruptField(
             views = @View(title = "负责人", sortable = true),
             edit = @Edit(title = "负责人", type = EditType.TAGS,
-                    tagsType = @TagsType(fetchHandler = CloudNode.class), notNull = true)
+                    tagsType = @TagsType(fetchHandler = CloudNodeProcess.class), notNull = true)
     )
     private String duty;
 
@@ -148,75 +129,5 @@ public class CloudNode extends MetaModelUpdateVo implements DataProxy<CloudNode>
     )
     private String remark;
 
-    @Transient
-    @Resource
-    private NodeManager nodeManager;
-
-    @Transient
-    @Resource
-    private EruptDao eruptDao;
-
-    @Override
-    public void afterUpdate(CloudNode cloudNode) {
-        DataProxy.super.afterUpdate(cloudNode);
-    }
-
-    @Override
-    public void beforeAdd(CloudNode cloudNode) {
-        if (null == cloudNode.getAccessToken()) cloudNode.setAccessToken(Erupts.generateCode(16).toUpperCase());
-    }
-
-    @Override
-    public void beforeUpdate(CloudNode cloudNode) {
-        this.beforeAdd(cloudNode);
-    }
-
-    @Override
-    public void afterFetch(Collection<Map<String, Object>> list) {
-        for (Map<String, Object> map : list) {
-            Optional.ofNullable(map.get(ACCESS_TOKEN)).ifPresent(it -> {
-                String token = it.toString();
-                map.put(ACCESS_TOKEN, token.substring(0, 3) + "******" + token.substring(token.length() - 3));
-            });
-            map.put(LambdaSee.field(CloudNode::getEruptNum), '-');
-            map.put(LambdaSee.field(CloudNode::getInstanceNum), '-');
-            map.put(LambdaSee.field(CloudNode::getVersion), '-');
-            map.put(LambdaSee.field(CloudNode::getEruptModuleNum), '-');
-            try {
-                MetaNode metaNode = nodeManager.getNode(map.get(NODE_NAME).toString());
-                Optional.ofNullable(nodeManager.getNode(map.get(NODE_NAME).toString())).ifPresent(metaNode1 -> {
-                    Function<Collection<String>, Object> function = (it) -> null == it ? 0 : String.format("<a href='javascript:alert(`%s`);'>%d</a>", String.join("\\u000a", it), it.size());
-                    map.put(LambdaSee.field(CloudNode::getEruptNum), function.apply(metaNode.getErupts()));
-                    map.put(LambdaSee.field(CloudNode::getInstanceNum), metaNode.getLocations().size());
-                    map.put(LambdaSee.field(CloudNode::getEruptModuleNum), function.apply(metaNode.getEruptModules()));
-                    map.put(LambdaSee.field(CloudNode::getVersion), metaNode.getVersion());
-                });
-            } catch (Exception e) {
-                map.put(LambdaSee.field(CloudNode::getVersion), String.format("<span style='color:#f00'>%s</span>", e.getMessage()));
-                log.warn("node warn → " + map.get(NODE_NAME), e);
-            }
-        }
-    }
-
-    @Override
-    public void afterDelete(CloudNode cloudNode) {
-        nodeManager.removeNode(cloudNode.getNodeName());
-    }
-
-    @Override
-    public List<String> fetchTags(String[] params) {
-        return eruptDao.getJdbcTemplate().queryForList("select name from e_upms_user", String.class);
-    }
-
-    @Override
-    public void bindTplData(Map<String, Object> binding, String[] params) {
-        CloudNode cloudNode = (CloudNode) binding.get(EngineConst.INJECT_ROW);
-        MetaNode metaNode = nodeManager.getNode(cloudNode.getNodeName());
-        if (null == metaNode) {
-            binding.put("instances", "[]");
-        } else {
-            binding.put("instances", GsonFactory.getGson().toJson(metaNode.getLocations()));
-        }
-    }
 
 }
